backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.0
sqlalchemy==2.0.30
python-multipart==0.0.9
pydantic==2.7.1
python-docx==1.1.2
pdfplumber==0.11.0